+++ NEWS.md
@@ -9,6 +9,7 @@
   * `bs_get_actor_lists()` retrieves all lists made by an actor
   * `bs_get_list()` retrieves a view of a list
 * Adds support for emoji in the text of posts, powered by the emoji package. (#11)
+* Adds `bs_url_to_uri()` to convert a URL to a `bskyr` URI.
 
 # bskyr 0.1.3